# refactor!: optimize clones and collections