--- a/quil-rs/src/program/calibration.rs
+++ b/quil-rs/src/program/calibration.rs
@@ -169,7 +169,8 @@ impl CalibrationSet {
         };
 
         // Add this instruction to the breadcrumb trail before recursion
-        let mut downstream_previous_calibrations = Vec::with_capacity(previous_calibrations.len() + 1);
+        let mut downstream_previous_calibrations =
+            Vec::with_capacity(previous_calibrations.len() + 1);
         downstream_previous_calibrations.push(instruction.clone());
         downstream_previous_calibrations.extend_from_slice(previous_calibrations);