Another set of optimizations, focused on searching for patterns in the code (e.g. cloning, creating a `Vec`) and seeing if it can be easily optimized in some way.

Each commit was benchmarked separately using an internal project, where any changes with negative results were discarded. I then ran benchmarks for both this project and the internal one, comparing `main` with this branch. The following summarizes the difference between `main` and these changes:

- `ScheduledProgram::from_program` saw a small performance increase overall, roughly 1-3%
- The `parser` saw little change in performance. Notably, only two example programs were called out as significant by `cargo bench`:
    - `empty-quil.quil` had a regression of 2.9% (~0.04 µs)
    - `good-include.quil` had a regression of 16% (~1.5 µs)
- `get_frames_for_instruction` seems noisy, with changes of 2-5% in either direction depending on the file.
    - `cargo bench` lists each benchmark as having 15-20% outliers, which I believe are left out of the statistical analysis
    - There appeared to be a few more improvements than regressions, though the noise limits how much this means anything
    - Outliers:
        - `good-classical-binaries.quil` had an improvement of 14.4% (~7.4 µs)
        - `good-pragmas.quil` had an improvement of ~7% (~0.7 µs)
        - `good-simple-params.quil` had a regression of ~2% (`include_blocked=false`) and ~8% (`include_blocked=true`), amounting to a change of ~0.3 µs and ~2 µs, respectively

The internal project has three main benchmarks. Two saw little change, while the third saw improvements across the board, ranging from 3%/22 µs to 14%/4ms.